cli/scripts: Updates flux install & Various fixes and improvements

+++ cli/scripts/src/cloud/gcp/setup/tekton/sf.sh
@@ -63,21 +63,31 @@ check_context "${cluster_config_context}"
 log_info "Generating sf secrets..."
 
 secretName="nexus"
+if [[ -f "${keys_path}/.${secretName}" ]]; then
 secret=$(kubectl create secret generic ${secretName} --dry-run=client -o yaml --from-env-file=${keys_path}/.${secretName}  | yq '.metadata |= {"name": "'"${secretName}"'", "namespace": "secrets", "annotations" : {"replicator.v1.mittwald.de/replication-allowed": "true", "replicator.v1.mittwald.de/replication-allowed-namespaces": "*"}}' -)
 echo "${secret}"  > ${secrets_path}/manifests/${secretName}.yaml
+fi
 
 secretName="npm-reader"
+if [[ -f "${keys_path}/.${secretName}" ]]; then
 secret=$(kubectl create secret generic ${secretName} --dry-run=client -o yaml --from-file=.nmprc=${keys_path}/.${secretName}  | yq '.metadata |= {"name": "'"${secretName}"'", "namespace": "secrets", "annotations" : {"replicator.v1.mittwald.de/replication-allowed": "true", "replicator.v1.mittwald.de/replication-allowed-namespaces": "*"}}' -)
 echo "${secret}"  > ${secrets_path}/manifests/${secretName}.yaml
+fi
 
 
 secretName="npm-publisher"
+if [[ -f "${keys_path}/.${secretName}" ]]; then
 secret=$(kubectl create secret generic ${secretName} --dry-run=client -o yaml --from-file=.nmprc=${keys_path}/.${secretName}  | yq '.metadata |= {"name": "'"${secretName}"'", "namespace": "secrets", "annotations" : {"replicator.v1.mittwald.de/replication-allowed": "true", "replicator.v1.mittwald.de/replication-allowed-namespaces": "*"}}' -)
 echo "${secret}"  > ${secrets_path}/manifests/${secretName}.yaml
+fi
 
 
 secretName="slack-webhook"
+if [[ -f "${keys_path}/.${secretName}" ]]; then
 secret=$(kubectl create secret generic ${secretName} --dry-run=client -o yaml --from-env-file=${keys_path}/.${secretName}  | yq '.metadata |= {"name": "'"${secretName}"'", "namespace": "secrets", "annotations" : {"replicator.v1.mittwald.de/replication-allowed": "true", "replicator.v1.mittwald.de/replication-allowed-namespaces": "*"}}' -)
 echo "${secret}"  > ${secrets_path}/manifests/${secretName}.yaml
+fi
 
 log_info "Done Generating sf secrets!"
+
+fi
